Split metrics aggregator into its own file
